+++ lib/src/models/event.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 class VietEvent {
   final DateTime date;
   final String? title;
@@ -11,6 +13,47 @@ class VietEvent {
     this.description,
     this.location,
   });
+
+  VietEvent copyWith({
+    DateTime? date,
+    String? title,
+    String? description,
+    String? location,
+    int? id,
+  }) {
+    return VietEvent(
+      date: date ?? this.date,
+      title: title ?? this.title,
+      description: description ?? this.description,
+      location: location ?? this.location,
+      id: id ?? this.id,
+    );
+  }
+
+  Map<String, dynamic> toMap() {
+    return {
+      'date': date.millisecondsSinceEpoch,
+      'title': title,
+      'description': description,
+      'location': location,
+      'id': id,
+    };
+  }
+
+  factory VietEvent.fromMap(Map<String, dynamic> map) {
+    return VietEvent(
+      date: DateTime.fromMillisecondsSinceEpoch(map['date']),
+      title: map['title'],
+      description: map['description'],
+      location: map['location'],
+      id: map['id']?.toInt(),
+    );
+  }
+
+  String toJson() => json.encode(toMap());
+
+  factory VietEvent.fromJson(String source) =>
+      VietEvent.fromMap(json.decode(source));
 
   @override
   bool operator ==(dynamic other) {