+++ lib/src/models/event_list.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 class VietEventList<T> {
   Map<DateTime, List<T>> events;
 
@@ -39,4 +41,21 @@ class VietEventList<T> {
   List<T> getEvents(DateTime date) {
     return events[date] ?? [];
   }
+
+  Map<String, dynamic> toMap() {
+    return {
+      'events': events,
+    };
+  }
+
+  factory VietEventList.fromMap(Map<String, dynamic> map) {
+    return VietEventList<T>(
+      events: Map<DateTime, List<T>>.from(map['events']),
+    );
+  }
+
+  String toJson() => json.encode(toMap());
+
+  factory VietEventList.fromJson(String source) =>
+      VietEventList.fromMap(json.decode(source));
 }